improve: allow l1-execution to fail in simulation
