+++ src/dataworker/index.ts
@@ -15,6 +15,7 @@ import {
   getRedisCache,
   waitForPubSub,
   averageBlockTime,
+  getChallengeRemaining,
 } from "../utils";
 import { spokePoolClientsToProviders } from "../common";
 import { Dataworker } from "./Dataworker";
@@ -79,16 +80,6 @@ function resolvePersonality(config: DataworkerConfig): string {
   }
 
   return "Dataworker"; // unknown
-}
-
-async function getChallengeRemaining(chainId: number): Promise<number> {
-  const provider = await getProvider(chainId);
-  const hubPool = getDeployedContract("HubPool", chainId).connect(provider);
-
-  const [proposal, currentTime] = await Promise.all([hubPool.rootBundleProposal(), hubPool.getCurrentTime()]);
-  const { challengePeriodEndTimestamp } = proposal;
-
-  return Math.max(challengePeriodEndTimestamp - currentTime, 0);
 }
 
 async function canProposeRootBundle(chainId: number): Promise<boolean> {
